+++ packages/components/src/Checkbox/Checkbox.test.tsx
@@ -2,21 +2,24 @@ import React from "react";
 import renderer from "react-test-renderer";
 import { cleanup, fireEvent, render } from "@testing-library/react";
 import { Checkbox } from ".";
+import { Text } from "../Text";
 
 afterEach(cleanup);
 
 it("renders a Checkbox", () => {
-  const tree = renderer
-    .create(<Checkbox label="Send me spam?" name="send_me_span" value="spam" />)
-    .toJSON();
-  expect(tree).toMatchSnapshot();
+  const { getByRole } = render(
+    <Checkbox label="Send me spam?" name="send_me_span" value="spam" />,
+  );
+  const checkbox = getByRole("checkbox");
+  expect(checkbox).toBeInTheDocument();
 });
 
 it("renders a disabled Checkbox", () => {
-  const tree = renderer
-    .create(<Checkbox label="Dont click me" disabled={true} />)
-    .toJSON();
-  expect(tree).toMatchSnapshot();
+  const { getByRole } = render(
+    <Checkbox label="Dont click me" disabled={true} />,
+  );
+  const checkbox = getByRole("checkbox");
+  expect(checkbox).toBeDisabled();
 });
 
 it("renders each variation of checked, defaultChecked and indeterminate", () => {
@@ -36,6 +39,35 @@ it("renders each variation of checked, defaultChecked and indeterminate", () => 
     expect(
       renderer.create(<Checkbox label="Foo" {...variation} />).toJSON(),
     ).toMatchSnapshot();
+  });
+});
+
+it("renders a description when set", () => {
+  const { getByText } = render(<Checkbox description="Checkers" />);
+  const description = getByText("Checkers");
+  expect(description).toBeInstanceOf(HTMLParagraphElement);
+});
+
+describe("With children components", () => {
+  it("renders a checkbox with children", () => {
+    const { getByText } = render(<Checkbox>{<Text>Content</Text>}</Checkbox>);
+    expect(getByText("Content")).toBeInTheDocument();
+  });
+
+  it("should still fire the onClick within the children", () => {
+    const handleLinkClick = jest.fn();
+    const { getByText } = render(
+      <Checkbox description="Checkers">
+        <Text>
+          I agree to the
+          <a onClick={handleLinkClick}>Terms of Service</a>
+        </Text>
+      </Checkbox>,
+    );
+    const TOSAnchorElement = getByText("Terms of Service");
+    expect(TOSAnchorElement).toBeInTheDocument();
+    fireEvent.click(TOSAnchorElement);
+    expect(handleLinkClick).toHaveBeenCalled();
   });
 });
 
@@ -61,10 +93,4 @@ describe("Clicking the checkbox it should call the handler", () => {
     fireEvent.click(getByLabelText("foo"));
     expect(clickHandler).toHaveBeenCalledWith(false);
   });
-});
-
-test("should render a description when set", () => {
-  const { getByText } = render(<Checkbox description="Checkers" />);
-  const description = getByText("Checkers");
-  expect(description).toBeInstanceOf(HTMLParagraphElement);
 });