+++ packages/components/src/DatePicker/DatePicker.css.d.ts
@@ -1,0 +1,10 @@
+declare const styles: {
+  readonly "datePickerWrapper": string;
+  readonly "fullWidth": string;
+  readonly "datePicker": string;
+  readonly "inline": string;
+  readonly "header": string;
+  readonly "month": string;
+};
+export = styles;
+