+++ packages/components/src/FormField/FormField.tsx
@@ -7,7 +7,7 @@ import React, {
   useImperativeHandle,
   useState,
 } from "react";
-import uuid from "uuid";
+import { v1 as uuidv1 } from "uuid";
 import { Controller, useForm, useFormContext } from "react-hook-form";
 import { FormFieldProps } from "./FormFieldTypes";
 import styles from "./FormField.css";
@@ -20,8 +20,10 @@ export function FormField(props: FormFieldProps) {
     autocomplete = true,
     children,
     defaultValue,
+    description,
     disabled,
     inputRef,
+    inline,
     keyboard,
     max,
     maxLength,
@@ -45,8 +47,8 @@ export function FormField(props: FormFieldProps) {
       ? useFormContext()
       : useForm({ mode: "onTouched" });
 
-  const [identifier] = useState(uuid.v1());
-
+  const [identifier] = useState(uuidv1());
+  const [descriptionIdentifier] = useState(`descriptionUUID--${uuidv1()}`);
   /**
    * Generate a name if one is not supplied, this is the name
    * that will be used for react-hook-form and not neccessarily
@@ -92,6 +94,8 @@ export function FormField(props: FormFieldProps) {
           readOnly: readonly,
           inputMode: keyboard,
           onChange: handleChange,
+          ...(description &&
+            !inline && { "aria-describedby": descriptionIdentifier }),
         };
 
         const textFieldProps = {
@@ -107,6 +111,7 @@ export function FormField(props: FormFieldProps) {
             value={rest.value}
             error={error}
             identifier={identifier}
+            descriptionIdentifier={descriptionIdentifier}
           >
             {renderField()}
           </FormFieldWrapper>
@@ -159,7 +164,7 @@ export function FormField(props: FormFieldProps) {
             newValue = parseFloat(newValue);
           }
 
-          onChange && onChange(newValue);
+          onChange && onChange(newValue, event);
           onControllerChange(event);
         }