+++ packages/components/src/FormField/FormFieldDescription.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { Text } from "../Text";
+
+interface FormFieldDescriptionProps {
+  readonly id: string;
+  readonly description: string;
+}
+
+export function FormFieldDescription({
+  id,
+  description,
+}: FormFieldDescriptionProps) {
+  return (
+    <div id={id}>
+      <Text variation="subdued">{description}</Text>
+    </div>
+  );
+}