+++ packages/components/src/FormField/FormFieldWrapper.tsx
@@ -1,4 +1,5 @@
 import React, {
+  Fragment,
   PropsWithChildren,
   RefObject,
   useEffect,
@@ -9,11 +10,14 @@ import classnames from "classnames";
 import { FormFieldProps } from "./FormFieldTypes";
 import styles from "./FormField.css";
 import { AffixIcon, AffixLabel } from "./FormFieldAffix";
+import { FormFieldDescription } from "./FormFieldDescription";
 import { InputValidation } from "../InputValidation";
+import { Content } from "../Content";
 
 interface FormFieldWrapperProps extends FormFieldProps {
   error: string;
   identifier: string;
+  descriptionIdentifier: string;
 }
 
 interface LabelPadding {
@@ -23,6 +27,8 @@ interface LabelPadding {
 
 export function FormFieldWrapper({
   align,
+  description,
+  descriptionIdentifier,
   placeholder,
   value,
   children,
@@ -72,8 +78,10 @@ export function FormFieldWrapper({
     setLabelStyle(getAffixPaddding);
   }, [value]);
 
+  const Tag = inline ? Fragment : Content;
+
   return (
-    <>
+    <Tag {...(!inline && { spacing: "small" })}>
       <div className={wrapperClasses} style={wrapperInlineStyle}>
         {prefix?.icon && <AffixIcon {...prefix} size={size} />}
         <div className={styles.inputWrapper}>
@@ -96,8 +104,18 @@ export function FormFieldWrapper({
           <AffixIcon {...suffix} variation="suffix" size={size} />
         )}
       </div>
-      {error && !inline && <InputValidation message={error} />}
-    </>
+      {!inline && (description || error) && (
+        <div>
+          {description && (
+            <FormFieldDescription
+              id={descriptionIdentifier}
+              description={description}
+            />
+          )}
+          {error && <InputValidation message={error} />}
+        </div>
+      )}
+    </Tag>
   );
 
   function getAffixPaddding() {