+++ packages/components/src/InputNumber/InputNumber.test.tsx
@@ -9,31 +9,35 @@ it("renders an input type number", () => {
   const tree = renderer.create(<InputNumber value={123} />).toJSON();
   expect(tree).toMatchInlineSnapshot(`
     <div
-      className="wrapper"
-      style={
-        Object {
-          "--formField-maxLength": undefined,
-        }
-      }
+      className="padded small"
     >
       <div
-        className="inputWrapper"
+        className="wrapper"
+        style={
+          Object {
+            "--formField-maxLength": undefined,
+          }
+        }
       >
-        <label
-          className="label"
-          htmlFor="123e4567-e89b-12d3-a456-426655440001"
-        />
-        <input
-          className="input"
-          id="123e4567-e89b-12d3-a456-426655440001"
-          name="generatedName--123e4567-e89b-12d3-a456-426655440001"
-          onBlur={[Function]}
-          onChange={[Function]}
-          onFocus={[Function]}
-          onKeyDown={[Function]}
-          type="number"
-          value={123}
-        />
+        <div
+          className="inputWrapper"
+        >
+          <label
+            className="label"
+            htmlFor="123e4567-e89b-12d3-a456-426655440001"
+          />
+          <input
+            className="input"
+            id="123e4567-e89b-12d3-a456-426655440001"
+            name="generatedName--123e4567-e89b-12d3-a456-426655440001"
+            onBlur={[Function]}
+            onChange={[Function]}
+            onFocus={[Function]}
+            onKeyDown={[Function]}
+            type="number"
+            value={123}
+          />
+        </div>
       </div>
     </div>
   `);