+++ packages/components/src/InputText/InputText.test.tsx
@@ -10,32 +10,36 @@ it("renders a regular input for text and numbers", () => {
     .toJSON();
   expect(tree).toMatchInlineSnapshot(`
     <div
-      className="wrapper"
-      style={
-        Object {
-          "--formField-maxLength": undefined,
-        }
-      }
+      className="padded small"
     >
       <div
-        className="inputWrapper"
+        className="wrapper"
+        style={
+          Object {
+            "--formField-maxLength": undefined,
+          }
+        }
       >
-        <label
-          className="label"
-          htmlFor="123e4567-e89b-12d3-a456-426655440001"
+        <div
+          className="inputWrapper"
         >
-          Favourite colour
-        </label>
-        <input
-          className="input"
-          id="123e4567-e89b-12d3-a456-426655440001"
-          onBlur={[Function]}
-          onChange={[Function]}
-          onFocus={[Function]}
-          onKeyDown={[Function]}
-          type="text"
-          value=""
-        />
+          <label
+            className="label"
+            htmlFor="123e4567-e89b-12d3-a456-426655440001"
+          >
+            Favourite colour
+          </label>
+          <input
+            className="input"
+            id="123e4567-e89b-12d3-a456-426655440001"
+            onBlur={[Function]}
+            onChange={[Function]}
+            onFocus={[Function]}
+            onKeyDown={[Function]}
+            type="text"
+            value=""
+          />
+        </div>
       </div>
     </div>
   `);
@@ -49,32 +53,36 @@ it("renders a textarea", () => {
     .toJSON();
   expect(tree).toMatchInlineSnapshot(`
     <div
-      className="wrapper textarea"
-      style={
-        Object {
-          "--formField-maxLength": undefined,
-        }
-      }
+      className="padded small"
     >
       <div
-        className="inputWrapper"
+        className="wrapper textarea"
+        style={
+          Object {
+            "--formField-maxLength": undefined,
+          }
+        }
       >
-        <label
-          className="label"
-          htmlFor="123e4567-e89b-12d3-a456-426655440002"
+        <div
+          className="inputWrapper"
         >
-          Describe your favourite colour?
-        </label>
-        <textarea
-          className="input"
-          id="123e4567-e89b-12d3-a456-426655440002"
-          onBlur={[Function]}
-          onChange={[Function]}
-          onFocus={[Function]}
-          onKeyDown={[Function]}
-          rows={3}
-          value=""
-        />
+          <label
+            className="label"
+            htmlFor="123e4567-e89b-12d3-a456-426655440003"
+          >
+            Describe your favourite colour?
+          </label>
+          <textarea
+            className="input"
+            id="123e4567-e89b-12d3-a456-426655440003"
+            onBlur={[Function]}
+            onChange={[Function]}
+            onFocus={[Function]}
+            onKeyDown={[Function]}
+            rows={3}
+            value=""
+          />
+        </div>
       </div>
     </div>
   `);
@@ -92,32 +100,36 @@ it("renders a textarea with 4 rows", () => {
     .toJSON();
   expect(tree).toMatchInlineSnapshot(`
     <div
-      className="wrapper textarea"
-      style={
-        Object {
-          "--formField-maxLength": undefined,
-        }
-      }
+      className="padded small"
     >
       <div
-        className="inputWrapper"
+        className="wrapper textarea"
+        style={
+          Object {
+            "--formField-maxLength": undefined,
+          }
+        }
       >
-        <label
-          className="label"
-          htmlFor="123e4567-e89b-12d3-a456-426655440003"
+        <div
+          className="inputWrapper"
         >
-          Describe your favourite colour?
-        </label>
-        <textarea
-          className="input"
-          id="123e4567-e89b-12d3-a456-426655440003"
-          onBlur={[Function]}
-          onChange={[Function]}
-          onFocus={[Function]}
-          onKeyDown={[Function]}
-          rows={4}
-          value=""
-        />
+          <label
+            className="label"
+            htmlFor="123e4567-e89b-12d3-a456-426655440005"
+          >
+            Describe your favourite colour?
+          </label>
+          <textarea
+            className="input"
+            id="123e4567-e89b-12d3-a456-426655440005"
+            onBlur={[Function]}
+            onChange={[Function]}
+            onFocus={[Function]}
+            onKeyDown={[Function]}
+            rows={4}
+            value=""
+          />
+        </div>
       </div>
     </div>
   `);