+++ packages/components/src/Menu/Menu.tsx
@@ -8,7 +8,7 @@ import React, {
   useRef,
   useState,
 } from "react";
-import uuid from "uuid";
+import { v1 as uuidv1 } from "uuid";
 import classnames from "classnames";
 import { AnimatePresence, motion } from "framer-motion";
 import { IconNames } from "@jobber/design";
@@ -62,13 +62,14 @@ export interface SectionProps {
 // eslint-disable-next-line max-statements
 export function Menu({ activator, items }: MenuProps) {
   const [visible, setVisible] = useState(false);
+  const fullWidth = activator?.props?.fullWidth || false;
   const [position, setPosition] = useState<Position>({
     vertical: "below",
     horizontal: "right",
   });
   const wrapperRef = createRef<HTMLDivElement>();
-  const buttonID = uuid();
-  const menuID = uuid();
+  const buttonID = uuidv1();
+  const menuID = uuidv1();
 
   useOnKeyDown(handleKeyboardShortcut, ["Escape"]);
   useLayoutEffect(() => {
@@ -90,7 +91,7 @@ export function Menu({ activator, items }: MenuProps) {
 
       setPosition(newPosition);
     }
-  }, [visible]);
+  }, [visible, fullWidth]);
 
   if (!activator) {
     activator = (
@@ -111,8 +112,12 @@ export function Menu({ activator, items }: MenuProps) {
     position.horizontal === "right" && styles.right,
   );
 
+  const wrapperClasses = classnames(styles.wrapper, {
+    [styles.fullWidth]: fullWidth,
+  });
+
   return (
-    <div className={styles.wrapper} ref={wrapperRef}>
+    <div className={wrapperClasses} ref={wrapperRef}>
       {React.cloneElement(activator, {
         onClick: toggle(activator.props.onClick),
         id: buttonID,