+++ packages/components/src/RadioGroup/RadioGroup.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from "react";
-import uuid from "uuid";
+import { v1 as uuidv1 } from "uuid";
 import { InternalRadioOption } from "./RadioOption";
 import styles from "./RadioGroup.css";
 
@@ -19,6 +19,11 @@ interface RadioGroupProps {
   onChange(newValue: string | number): void;
 
   /**
+   * Defines the aria label that describes the radio group.
+   */
+  readonly ariaLabel: string;
+
+  /**
    * The name of the radio group, that links the radio options back up
    * to the group.
    *
@@ -30,11 +35,12 @@ interface RadioGroupProps {
 export function RadioGroup({
   children,
   value,
+  ariaLabel,
   onChange,
-  name = uuid.v1(),
+  name = uuidv1(),
 }: RadioGroupProps) {
   return (
-    <div role="radiogroup" className={styles.radioGroup}>
+    <div role="radiogroup" aria-label={ariaLabel} className={styles.radioGroup}>
       {React.Children.map(children, option => (
         <InternalRadioOption
           checked={value === option.props.value}