+++ packages/components/src/RecurringSelect/DateCellCheckbox.css.d.ts
@@ -1,0 +1,7 @@
+declare const styles: {
+  readonly "checkboxWrapper": string;
+  readonly "dayCheckboxInput": string;
+  readonly "dayCheckbox": string;
+};
+export = styles;
+