+++ packages/components/src/RecurringSelect/RecurringSelect.css.d.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly "container": string;
+  readonly "fullWidthWrapper": string;
+};
+export = styles;
+