+++ packages/components/src/RecurringSelect/components/index.tsx
@@ -1,0 +1,4 @@
+export { WeeklySelect } from "./WeeklySelect/WeeklySelect";
+export { MonthlySelect } from "./MonthlySelect/MonthlySelect";
+export { MonthlyDayOfWeekSelect } from "./MonthlyDayOfWeekSelect/MonthlyDayOfWeekSelect";
+export { DayOfMonthSelect } from "./DayOfMonthSelect/DayOfMonthSelect";