+++ packages/components/src/Table/Table.test.tsx
@@ -1,6 +1,15 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import { Cell, CellCurrency, CellNumeric, Footer, Header, Row, Table } from ".";
+import {
+  Body,
+  Cell,
+  CellCurrency,
+  CellNumeric,
+  Footer,
+  Header,
+  Row,
+  Table,
+} from ".";
 
 it("renders a table", () => {
   const tree = renderer
@@ -12,24 +21,26 @@ it("renders a table", () => {
           <Cell>Cost</Cell>
           <Cell>Crew</Cell>
         </Header>
-        <Row>
-          <Cell>Nostromo</Cell>
-          <Cell>Towing Vehicle</Cell>
-          <CellCurrency value={42000000} />
-          <CellNumeric value={7} />
-        </Row>
-        <Row>
-          <Cell>Rodger Young</Cell>
-          <Cell>Corvette Transport</Cell>
-          <Cell />
-          <CellNumeric value={200} />
-        </Row>
-        <Row>
-          <Cell>USS Enterprise</Cell>
-          <Cell>Constitution</Cell>
-          <Cell />
-          <CellNumeric value={205} />
-        </Row>
+        <Body>
+          <Row>
+            <Cell>Nostromo</Cell>
+            <Cell>Towing Vehicle</Cell>
+            <CellCurrency value={42000000} />
+            <CellNumeric value={7} />
+          </Row>
+          <Row>
+            <Cell>Rodger Young</Cell>
+            <Cell>Corvette Transport</Cell>
+            <Cell />
+            <CellNumeric value={200} />
+          </Row>
+          <Row>
+            <Cell>USS Enterprise</Cell>
+            <Cell>Constitution</Cell>
+            <Cell />
+            <CellNumeric value={205} />
+          </Row>
+        </Body>
         <Footer>
           <Cell />
           <Cell />