+++ packages/components/src/Typography/css/FontFamilies.css.d.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly "base": string;
+  readonly "display": string;
+};
+export = styles;
+