+++ packages/hooks/CHANGELOG.md
@@ -8,6 +8,25 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.1](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.9.0...@jobber/hooks@1.9.1) (2021-12-15)
+
+**Note:** Version bump only for package @jobber/hooks
+
+
+
+
+
+# [1.9.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.8.0...@jobber/hooks@1.9.0) (2021-10-29)
+
+
+### Features
+
+* **components:** Add description to base FormField component ([#714](https://github.com/GetJobber/atlantis/issues/714)) ([3fae72e](https://github.com/GetJobber/atlantis/commit/3fae72e9f483ece23263c9b2c981060248223db5))
+
+
+
+
+
 # [1.8.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@1.7.2...@jobber/hooks@1.8.0) (2021-10-13)